Money types: convert Sales domain Double -> BigDecimal (slice 5/N)
The big, GL-touching slice. Test-first discipline: baselined PaymentService
settlement math (5 cases) and confirmed PostingEngineContractTest (37) as the
GL safety net BEFORE flipping, then kept both green after.

Entities flipped (money -> BigDecimal @Column(precision=15,scale=2); percentage
rate fields kept as Double):
- SalesInvoice (subTotal/taxTotal/invoiceTotal/amountPaid/balance/
  billDiscountAmount/deliveryCharge/roundOff/creditLimit; billDiscount=rate)
- SalesInvoiceItem (price/cost/footerDiscount/taxAmount/grossAmount/netAmount;
  discount,taxRate=rates)
- SalesOrder, SalesOrderItem, Payment, SalesReturn, SalesReturnItem,
  DeliveryNoteItem (price/cost; disc,tax=rates), DeliveryNoteItemRequest

Service math rewritten in BigDecimal, semantics preserved:
- SalesInvoiceService.finalizeInvoiceTotals: per-line accumulation, HALF_UP
  rounding, % bill-discount derivation, total/balance. The old
  `paid > total + 0.0001d` float epsilon becomes an exact compareTo (intended
  behaviour change — documented in SalesInvoiceTotalsTest).
- SalesOrderService totals + advance/RV reconciliation; PaymentService
  recomputeInvoiceBalances running fold; ReceiptVoucherService settlement guards
  and syncLinkedInvoice/resolveInvoiceStatus.

Consumers fixed: PostingEngineService (dropped now-redundant BigDecimal.valueOf
wraps on invoice/return getters, signum() guards, added nz(BigDecimal) overload),
DeliveryNoteService (COGS/revenue recognition), DashboardService/PosLookupService/
PosCheckoutController/FinancialReportService (bridged at DTO boundaries).

Caught + fixed a regression the Mockito unit tests miss but contextLoads
surfaced: JPQL constructor projections and aggregate return types over the
flipped columns.
- StatementEntryDTO / PriceHistoryDTO projections: CAST(0 AS big_decimal) and
  PriceHistoryDTO.rate -> BigDecimal so the SELECT NEW constructor binds.
- ~13 `Double`-declared SUM() finders over now-NUMERIC columns would
  ClassCastException at runtime (JPQL SUM(bigdecimal)=BigDecimal). Wrapped each
  in CAST(... AS double) so the method return type and all callers are unchanged.

Verification: mvn -o test = 134 run; the ONLY failures are the pre-existing
WarehouseServiceTest (strict-stubbing, unrelated, proven on clean parent).
contextLoads now PASSES (full Spring context + every JPQL validated against a
live DB) — a stronger green than the baseline where it was the known-failing
@SpringBootTest. GL safety net (PostingEngineContractTest 37, PostingEngine 5),
SalesInvoiceTotalsTest 10, PaymentServiceTest 5, ReceiptVoucherServiceTest 2 all
green.

+++ billbull-backend/src/main/java/com/billbull/backend/dashboard/DashboardService.java
@@ -204,7 +204,7 @@ public class DashboardService {
                     inv.getId().toString(),
                     inv.getInvoiceNumber(),
                     inv.getCustomerName(),
-                    inv.getInvoiceTotal() != null ? inv.getInvoiceTotal() : 0d,
+                    inv.getInvoiceTotal() != null ? inv.getInvoiceTotal().doubleValue() : 0d,
                     inv.getStatus() != null ? inv.getStatus().toString() : "DRAFT",
                     inv.getInvoiceDate() != null ? inv.getInvoiceDate().format(fmt) : null,
                     createdAtStr
